requirements.txt
chat_exporter==2.2.1
discord.py==2.4.0
fastapi==0.115.12
motor==3.0.0
pymongo==4.2.0
python_dateutil==2.8.2
uvicorn==0.18.1
